tests/stream/helpers/requirements.txt
pytest==6.2.5
requests>=2.26.0
pyOpenSSL==23.0.0
pytest-cases==3.6.5
unidiff==0.7.0
docker-compose==1.29.2
boto3==1.20.4
urllib3==1.26.7
pytest-html==3.1.1
psycopg2-binary==2.9.3
kafka-python==2.0.2
timeplus==1.1.2
sseclient-py>=1.7.2
proton-driver>=0.2.4
